+++ src/Fantomas.Core.Tests/MultilineBlockBracketsOnSameColumnArrayOrListTests.fs
@@ -105,7 +105,7 @@ let ``array comprehensions`` () =
         """
 let a1 = [| 0 .. 99 |]
 let a2 = [| for n in 1 .. 100 do if isPrime n then yield n |]"""
-        config
+        { config with MaxIfThenShortWidth = 25 }
     |> prepend newline
     |> should
         equal